Force movie titles to be in English
It seems akas.imdb.com behavior changed and most titles were returned in
French. We now force to have English titles. This is not ideal for actual
french movies, but since the vast majority of movies are from English
speaking countries, it's the best (easy) solution so far.

So we basically reverted to previous behavior of mQueue.